requirements.txt
boto3>=1.42.34
fastapi>=0.128.0
langchain-core>=1.2.7
langchain-google-genai>=4.2.0
langgraph>=1.0.7
mcp>=1.25.0
requests>=2.32.5
uvicorn>=0.40.0
websockets>=15.0.1
psycopg2-binary>=2.9.9
